+++ core/src/db/meta/MetaSession.h
@@ -127,8 +127,18 @@ MetaSession::Select(const std::string& field, const std::vector<U>& values,
     for (auto raw : attrs) {
         auto resource = snapshot::CreateResPtr<T>();
         std::unordered_map<std::string, std::string>::iterator iter;
-        auto mf_p = std::dynamic_pointer_cast<snapshot::MappingsField>(resource);
-        if (mf_p != nullptr) {
+
+        if (auto mf_p = std::dynamic_pointer_cast<snapshot::FlushableMappingsField>(resource)) {
+            iter = raw.find(F_MAPPINGS);
+            if (iter != raw.end()) {
+                auto mapping_json = nlohmann::json::parse(iter->second);
+                std::set<int64_t> mappings;
+                for (auto& ele : mapping_json) {
+                    mappings.insert(ele.get<int64_t>());
+                }
+                mf_p->GetFlushIds() = mappings;
+            }
+        } else if (auto mf_p = std::dynamic_pointer_cast<snapshot::MappingsField>(resource)) {
             iter = raw.find(F_MAPPINGS);
             if (iter != raw.end()) {
                 auto mapping_json = nlohmann::json::parse(iter->second);