+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Please mark all changes in change log and use the issue from GitHub
 -   \#3652 Proto of C++ sdk is different from milvus server
 -   \#3668 Docker exit without any logs
 -   \#3672 0.11.0 docker image is 200M larger due to the un-expected installation of openblas
+-   \#3689 Client get stuck if drop collection during creating index
 
 ## Feature
 -   \#2319 Redo metadata to support MVCC